Moved functions to .cpp files
